upgrade to xxh3 hashing pkg
