+++ _example/go.mod
@@ -1,6 +1,8 @@
 module github.com/go-chi/httprate/_example
 
-go 1.22.5
+go 1.23.0
+
+toolchain go1.24.1
 
 replace github.com/go-chi/httprate => ../
 
@@ -9,4 +11,8 @@ require (
 	github.com/go-chi/httprate v0.0.0-00010101000000-000000000000
 )
 
-require github.com/cespare/xxhash/v2 v2.3.0 // indirect
+require (
+	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
+	github.com/zeebo/xxh3 v1.0.2 // indirect
+	golang.org/x/sys v0.30.0 // indirect
+)
