+++ go.mod
@@ -1,7 +1,14 @@
 module github.com/go-chi/httprate
 
-go 1.17
+go 1.23.0
 
-require github.com/cespare/xxhash/v2 v2.3.0
+toolchain go1.24.1
 
-require golang.org/x/sync v0.7.0 // indirect
+require github.com/zeebo/xxh3 v1.0.2
+
+require golang.org/x/sys v0.30.0 // indirect
+
+require (
+	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
+	golang.org/x/sync v0.12.0
+)
