+++ limit_key.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/cespare/xxhash/v2"
+	"github.com/zeebo/xxh3"
 )
 
 func LimitCounterKey(key string, window time.Time) uint64 {
-	h := xxhash.New()
+	h := xxh3.New()
 	h.WriteString(key)
 	h.WriteString(fmt.Sprintf("%d", window.Unix()))
 	return h.Sum64()
